practice_4/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-seed==0.3.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
Faker==37.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
uritemplate==4.1.1
